Switch dependency management to use Python 3 on rlogin
I wrote a script to install Python 3 locally and use this
environment over the Python 2 one available by default. This
way, we can all easily use the most up to date versions.

It install Python 3 to ~/local to resemble /usr/local and links
binaries to ~/bin to resemble /usr/bin. This makes it easy to
"rm -rf" the single directory to completely uninstall this
installation whenever you need to.